+++ imagedisplay.h
@@ -7,7 +7,7 @@ G_DECLARE_FINAL_TYPE( Imagedisplay, imagedisplay,
 	VIPSDISP, IMAGEDISPLAY, GtkDrawingArea )
 
 void imagedisplay_image_to_gtk( Imagedisplay *imagedisplay, 
-	int x_image, int y_image, double *x_gtk, double *y_gtk );
+	double x_image, double y_image, double *x_gtk, double *y_gtk );
 void imagedisplay_gtk_to_image( Imagedisplay *imagedisplay, 
 	double x_gtk, double y_gtk, double *x_image, double *y_image );
